Add converge_edge_sources function to the network functions
Add the converge_edge_sources function to help converging the sources of
the graphs forming a union to one attribute.
The problem rises when doing a union of two graphs with the source attribute
on its edges. Instead of putting edges with the source attribute correctly
defined in the unified graph, igraph splits into source_1 and source_2.

converge_edge_sources fixes this, and makes it so all edges on the unified
graph have only the one source attribute which is defined correctly