Creating an R project in the root directory
In this commit we create an R project in the repo's
root directory.
This enables git integration in RStudio, but more
importantly, it will allow us to use the `here`
package to make portable relative paths from the
project root in order to source files and access
data files instead of hard-coding a path relevant
to a specific machine.
The prerequisite for this to work is to open the project file so that RStudio will switch the working directory to the project root