Add dependency management using the renv package
In this commit we used the renv package to generate dependency management capabilities to the project.
This creates the renv.lock file to track all packages used in the project and their versions (similar to package.json of npm) and other auxiliary files to track the dependencies.
It also creates a local renv/library folder that contains the packages (or symbolic links to packages if they exist in the global package library). That library is gitignored.
See also:
https://ecorepsci.github.io/reproducible-science/renv.html
and
https://rstudio.github.io/renv/articles/renv.html