Replace hard coded paths with here packacge
In order to make the code in the repo portable and reproducable, the hardcoded paths had to be replaced.
Instead, the here package leverages the `Rproj` file in the root directory to locate the correct root path of the projects.
You can therefore supply the here function with strings representing the relative path from the root (each level a separate string).

Unfortunately, many files cannot benefit from this because of one of two reasons:
  1) They refer to paths that no longer exist
  2) They refer to paths outside of the
  repository.
Those files will need to be addressed in the future.
They include: everything in 'biomarker_selection/EDA', 'biomarker_selection/pre-processing', and the two 'olink_analysis' files in 'biomarker_selection',