Replace hard coded paths using file choice dialog
In this commit, we create the `define_datasets_root` function,
which checks if a variable named `datasets_root_directory` exists.
If so, it asks the user whether they want to redefine it.
If the user answered 'y', or the variable didn't exist,
the user is prompted to choose the directory which is the root of all
datasets.
We then replace the hardcoded paths with
`file.path(datasets_root_directory, "<PREV-PATH-WITHOUT-.../DATASETS/>")`
This solved the issue with paths outside of the repository refering to
datasets.
It doesn't solve the issue for path that do not exist.
We also delete two files that had the "-old" suffix,
and were basically old versions of other files.
We also remove some trailling whitespace