Separate more functions
In this commit we separate more functions.
We start with taking
`sapply(colnames(tab), function(n) strsplit(n, "_")[[1]][1])`
and refactor it to a function to be extracted to
`data_processing_functions`
WAIT!!! I meant to only take the anonymous function and extract it
to the other file.
That is: `strsplit(n, "_")[[1]][1]`.
Next, I will create `auxiliary_statistical_functions` to extract those
functions from `clean_kth`.
The first function I tackled was `kw_for_diagnoses`, where I wanted to
generalize it from the limited hardcoded usage of performing
Kruskal-Wallis test for all the Diagnoses in the given the hgnc symbol.
I wrote `kruskal_wallis_test_for_sample_groups`, where instead
of "Diagnosis" it accepts a symbol of a grouping feature column name,
instead of "HGNC_Symbol" it accepts a symbol of a diffetentiating
feature column name, instead of "median_ab_readout" it accepts a symbol
of a measurement column name. The next challenge is to generalize
`test_pairwise_wilcoxon`. Between the usage of `kw_for_diagnoses`
and the usage of `test_pairwise_wilcoxon` there are several things going
on which I am going to fold into my new implementation.
First up, it keeps only the rows where the adjusted p value
(the adjustment happens on the way) is below a significance value.
Next it creates all possible pairs of the grouping feature as a named
list(name being "\<feature1>_vs_\<feature2>") of vectors with the two
members of the pair (["\<feature1>", "\<feature2>"]).
Only then does it `sapply` it while also supplying the symbol of a
diffetentiating feature column name and a vector of all diffetentiating
feature values.
So the following auxiliary functions were created to assist with that:

- `is_adjusted_pvalue_significant_predicate` which adjusts the p values
  of a dataframe and then makes a vector predicate to be used on the
  original df
- `unique_table_column_values` which uses the symbol of the column name
  to extract it and calls `unique` on the result
- `collect_wilcoxon_pvalues_of_all_pairs_and_proteins` which does **2**
   layers of sapply:
  - It itself is called from a sapply on all the proteins and
  - Inside there is a call for every pair