[Experimental] Attempt to perform PCA on kth and visualize
In this commit, we take the numeric columns of kth dataset,
we impute missing values and perform PCA.

We then try to visualize with `biplot` but it doesn't look good.

We then scale the imputed kth and perform another pca, then add
the `Class` column from original db and try to use ggplot to
visualize it which is not working at the moment.

DISCLAIMER: The code is based on ChatGPT advice