[Experimental] Start with community analysis
In this WIP, we start analysing the communities of the unified graph
and also the vertex intersection graph (graph containing only vertices
that are in both original graphs).

We then repurpose a function to enrich pathways according to kegg
from the prioritisation script to perform enrichment on all communities
to mixed results.

Needs to be polished.