[Community-analysis] Fix enrichment and perform aggregation
In this commit we fix the `community_enrichent` function running into an issue in the `bind_rows` call (due to some bug causing a type mismatch) by pruning empty dataframes before that stage.
This pruning also happens to the list returned from the function.

In addition, we group by gene and aggregate so that all terms of a gene are concatenated with a ';' delimiter