[pruning-networks]: Extract get_all_interactions_that_have_backing_uniprots
In this commit, we extract the process of using the interactions excel
and the dictionary excel to have a df that has both network objects
and uniprots of target and source.
In this funtction, empty or invalid uniprot rows are filtered out, and
there is a bool controlling whether you end up with uniprots or
network object names.