[Network pruning]: Refine interaction processing
In this commit, we continue refining the interaction processing.
We expand the choice of whether to work with uniprots ore not to
the entire processing process. We decide at the beginning whether
`work_with_uniprots` is true. We then perform the analysis with that
in mind, including the interaction dfs containing uniprots/containing
network objects, and the vertices being/not being uniprots.
So now we have a `work_with_uniprots` boolean at the top of the file
that controls that.
We prepare the vertices differently for each case, and from there it
runs the same.
In fact, I will extract a function for that in order to have a cleaner
environment.
The new function, `extract_vertices_for_all_graphs`, extracts either
the uniprots or network objects (depending on boolean) and then returns
a list of lists of vertices of all future graphs.
We also create the auxiliary function called `named_list` that takes
items and a character vector of names and creates a named list as long
as they have the same length