[EDA]: Add function to meld together pieces of a fractured dataset
In this commit we add `meld_fractured_dataset`, which accepts 2 or 3
file paths, a named vector for the join column names (of length
\#datasets - 1) containing the mapping between the lower indexed
dataset and the higher one, and `select_columns` and `exclude_columns`
one of which to be used if not all columns are desired.
We also add the `read_df_from_file` that can read any of the
following types: tsv, csv, xls and xlsx. In the case of a password
protected xlsx it can also handle it by calling our other
function, `read_password_xlsx`