[EDA]: Add male vs female calculation for emif
In this commit, we use the workflow of:
`meld_fractured_dataset`->`melt`->`kruskal_wallis_test_for_sample_groups`
->`keep_only_significant_entries`->`test_pairwise_wilcoxon`
->`p.adjust`->turning into dataframe
on the emif dataset for the case of male vs female.

We also do the fix in this workflow for the case where we got a numeric
result from the pairwise wilcoxon.