[EDA]: Shift to usage of wrapper function and perform analysis on 3 datasets
In this commit, we shift the existing analysis to use the new
wrapper function for the kw-wilcoxon process.
We also add the missing olink dataset, and make sure we have gender,
age_group and diagnosis for all 3 datasets.
Unfortunately, the age_group one for olink does not work due to some bug