[BiomarkerDistribution]: Add biomarker distribution plotting
In this commit, we add the new `biomarker_distribution` file.
In this file, we take the raw data for emif, kth and olink.
We extract the overlapping proteins.
We then filter out non overlapping proteins, we make sure all
three filtered datasets use UniProts, we unify the naming of
the Diagnosis, we sort by age, we unify the naming of the columns
and we add an Id column that is just the row number.
We then bind the 3 filtered dfs (with the cohort column used
to recognize the original dataset of each row), we pivot long
so that we have a uniprot and a value per row.
We then find the means for each cohort+uniprot combination
and use it to normalize the values.
We finally plot the biomarker distribution.