[Pathway enrichment]: Perform additional pathway enrichment
In this commit, we wrap the community enrichment with an if else
that checks for an excel file with the data. If it doesn't exist,
you call EnrichR with a new wrapper function that has fault tolerance
(if it times out, you can stop it and it will try again from the correct
index). You then save it to the file. Else, you read the info from the
file.

After that, we compared the pathways with and without the communities
for the unified graph, by filtering out the ones that are in both.