[EDA]: Divide age groups according to quantiles and not arbitrarily
In this commit, we add a function to compute the breaks and labels
using the `quantile` function, and while it does not produce 3 perfectly
equal groups, it is better than arbitrarily dividing into groups.
In addition, I rearrange the `age_sex_pvalue_calculations` such that
the data processing of all datasets is at the top, then the kw and
wilcoxon and then the box plots.