[EDA]: Rename for uniformity and boxplot only significant proteins
In this commit, we do renamings of columns and values to have things
more uniform. We also do the supercharge melt (see variables of similar
name) to emif and olink.
We then do the kw-wilcoxon to all datasets according to
diagnosis-age-group and extract only the significant proteins from
those. We then do the boxplots only for those.
For KTH and Emif it helps. For Olink we are still left with 737
proteins.

Additionally, we do a small fix in the boxplot function