[EDA]: Create redux dataset comparisons and correct venn diagrams
In this commit, we use the results of the previous one and move from
there. We do a redux version of the `compare_datasets` function,
which also uses HGNC Symbols and not uniprots, to prepare for the
plotting of the venn diagrams which are then plotted for all the
proteins using a trick of reusing existing function just with a
significant value parameter of 1 (therefore not filtering proteins).
We add to the mix the adni dataset in this commit, but again without
cleaning it.

We end up with the correct numbers for the overlap.

The old `compare_datasets` script has been retained and renamed
into `compare_datasets_cleaned`.